Fixes cppcache  warnings on Windows.

+++ cppcache/include/geode/ResultSet.hpp
@@ -52,7 +52,7 @@ class _GEODE_EXPORT ResultSet : public SelectResults {
    *
    * @returns the number of items in the ResultSet.
    */
-  virtual int32_t size() const override = 0;
+  virtual size_t size() const override = 0;
 
   /**
    * Index operator to directly access an item in the ResultSet.
